+++ Sources/Atoms/Other.swift
@@ -1,0 +1,23 @@
+import Foundation
+
+// MARK: All atoms that don't have a specific implementation yet
+extension Atom {
+  class TRAK: Atom {}
+  class TKHD: Atom {}
+  class MDIA: Atom {}
+  class MDHD: Atom {}
+  class HDLR: Atom {}
+  class MINF: Atom {}
+  class VMHD: Atom {}
+  class DINF: Atom {}
+  class DREF: Atom {}
+  class STBL: Atom {}
+  class STSD: Atom {}
+  class STTS: Atom {}
+  class STSS: Atom {}
+  class STSC: Atom {}
+  class STSZ: Atom {}
+  class STCO: Atom {}
+  class UDTA: Atom {}
+  class Unknown: Atom {}
+}
